Refactor RandomUtil
